differ acquire/release behaviors between STL and ASTF mode
